refactor(D1): de-ghost the recommender subsystem — explicit composition
Implements deep revision D1 from docs/architecture/2026-07-full-repo-review.md:

- RecommenderService inherits ModelSelectionMixin directly; update_config /
  reset call _apply_model_settings / _reset_model_settings explicitly.
- Delete expert_backend/recommenders/_service_integration.py (import-time
  monkey-patching) and the shadowed legacy run_analysis_step2 generator on
  AnalysisMixin; the single model-aware generator now lives on AnalysisMixin,
  decomposed into _run_step2_discovery / _enrich_step2_results helpers.
- Fix the antenna_meta mirror-drift bug: the production step-2 generator
  dropped the field added in 2dd2ced, leaving the frontend AntennaNotice
  dead in production. Regression-guarded in test_model_composition.py.
- Rescue the orphaned root tests/ (8 files, 144 tests, zero CI coverage)
  into expert_backend/tests/, merging the test_recommenders_registry.py
  collision and rewriting test_service_integration.py as
  test_model_composition.py. The rescue caught a real bug: the overflow
  path filter's segment scan never matched substation names containing
  underscores (VL_LOOP) — fixed in _action_touches_path.
- Update both CLAUDE.md trees, docs/backend/*, README tree, benchmark
  patch target, CHANGELOG.

1052 passed / 6 skipped; ruff, mypy, code-quality gate green.
